+++ internal/service/category/tests/create_test.go
@@ -15,8 +15,8 @@ import (
 func TestCreate(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
-	id := int64(gofakeit.Int64())
-	domainID := int64(gofakeit.Int64())
+	id := gofakeit.Int64()
+	domainID := gofakeit.Int64()
 	req := &model.NewCategory{
 		Name:     gofakeit.Name(),
 		DomainID: domainID,
