+++ internal/service/question/tests/add_options_test.go
@@ -15,7 +15,7 @@ import (
 func TestAddOptions(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
-	id := int64(gofakeit.Int64())
+	id := gofakeit.Int64()
 	req := []*model.NewQuestionOption{
 		{
 			Text:      gofakeit.Question(),
